Typestack refactor

+++ server/src/api/Services/PartyService.ts
@@ -7,7 +7,6 @@ import { SpotifyOAuthState } from "../Types/general/SpotifyOAuthState";
 import { PartyDB } from "../Types/DatabaseMaps/PartyDB";
 import { PartyBuilder } from "../Factory/PartyBuilder";
 import { PartyDatabaseService } from "./Database/PartyDatabaseService";
-import { PartyHistoryDatabaseService } from "./Database/PartyHistoryDatabaseService";
 import { UserDatabaseService } from "./Database/UserDatabaseService";
 import { UserDB } from "../Types/DatabaseMaps/UserDB";
 import { UserBuilder } from "../Factory/UserBuilder";
@@ -22,6 +21,8 @@ import { PlaylistVoteEnum } from "../Types/Enums/PlaylistVoteEnum";
 import moment from "moment";
 import { PartyEvent } from "../Factory/PartyEventEmitterBuilder";
 import { EventEmitterService } from "./EventEmitterService";
+import { CronJobService } from "./CronJobService";
+import { userSort } from "./sorts";
 
 @Service()
 export class PartyService {
@@ -34,11 +35,20 @@ export class PartyService {
         private spotifyStateService: SpotifyStateService,
         private userDatabaseService: UserDatabaseService,
         private partyDatabaseService: PartyDatabaseService,
-        private partyHistoryDatabaseService: PartyHistoryDatabaseService,
         private playlistEntryDatabaseService: PlaylistEntryDatabaseService,
         private playlistVoteDatabaseService: PlaylistVoteDatabaseService,
         private upNextService: UpNextService,
+        private cronJobService: CronJobService
     ) {
+        this.cronJobService.newCronJob({
+            pattern: '*/10 * * * *',
+            method: () => {
+                const t24Hr = moment().subtract(24, 'hours').valueOf();
+                this.getAllParties()
+                    .filter(e => e.startTime < t24Hr)
+                    .forEach(e => this.removePartyByPartyId(e.id));
+            }
+        });
     }
 
     public async newParty(state: SpotifyOAuthState, code: string): Promise<any> {
@@ -81,10 +91,8 @@ export class PartyService {
     public removePartyByPartyId(partyId: string): void {
         this.upNextService.stopPartyByPartyId(partyId);
         this.spotifyStateService.stopSpotifyStateForParty(partyId);
-        this.partyHistoryDatabaseService.removeHistoryForParty(partyId);
-        this.playlistEntryDatabaseService.removePlaylistEntriesByPartyId(partyId);
-        this.userDatabaseService.removeAllUsersWithPartyId(partyId);
         this.partyDatabaseService.removePartyByPartyId(partyId);
+        this.eventEmitterService.emitEventAt(partyId, PartyEvent.PARTY_GONE);
     }
 
     public removeNewPartyEntry(partyId: string): void {
@@ -102,6 +110,7 @@ export class PartyService {
             .withTrackingId(token.trackingId)
             .build();
         this.userDatabaseService.insertUser(user);
+        this.emitUsersUpdate(token.partyId);
         return user.id;
     }
 
@@ -125,6 +134,7 @@ export class PartyService {
     }
 
     public upvoteSong(partyId: string, userId: string, playlistEntryId: string): void {
+        const entry = this.playlistEntryDatabaseService.getPlaylistEntryById(playlistEntryId);
         const userVotes = this.playlistVoteDatabaseService.getVotesForUserOnEntry(userId, playlistEntryId);
         if (userVotes.length === 0) {
             // perfect the user has yet to upvote
@@ -134,19 +144,26 @@ export class PartyService {
                 type: PlaylistVoteEnum.UPVOTE
             });
             this.playlistEntryDatabaseService.addUpVote(playlistEntryId);
+            this.userDatabaseService.updateUserScore(entry.addedBy, 1);
             this.emitPlaylistUpdate(partyId);
-        } else if (userVotes.length === 1) {
-            // user has made a vote, if its a downvote, get rid of it
-            if (userVotes[0].type === PlaylistVoteEnum.DOWNVOTE) {
-                this.playlistEntryDatabaseService.removeDownVote(playlistEntryId);
-                this.playlistEntryDatabaseService.addUpVote(playlistEntryId);
-                this.playlistVoteDatabaseService.updateVote(playlistEntryId, userId, PlaylistVoteEnum.UPVOTE);
-                this.emitPlaylistUpdate(partyId);
-            }
+            this.emitUsersUpdate(partyId);
+        } else if (userVotes.length === 1 && userVotes[0].type === PlaylistVoteEnum.DOWNVOTE) {
+            this.playlistEntryDatabaseService.removeDownVote(playlistEntryId);
+            this.playlistEntryDatabaseService.addUpVote(playlistEntryId);
+            this.playlistVoteDatabaseService.deleteVote(playlistEntryId, userId);
+            this.playlistVoteDatabaseService.insertVote({
+                type: PlaylistVoteEnum.UPVOTE,
+                playlistEntryId,
+                userId
+            });
+            this.userDatabaseService.updateUserScore(entry.addedBy, 1);
+            this.emitPlaylistUpdate(partyId);
+            this.emitUsersUpdate(partyId);
         }
     }
 
     public downvoteSong(partyId: string, userId: string, playlistEntryId: string): void {
+        const entry = this.playlistEntryDatabaseService.getPlaylistEntryById(playlistEntryId);
         const userVotes = this.playlistVoteDatabaseService.getVotesForUserOnEntry(userId, playlistEntryId);
         if (userVotes.length === 0) {
             // perfect the user has yet to upvote
@@ -156,19 +173,25 @@ export class PartyService {
                 type: PlaylistVoteEnum.DOWNVOTE
             });
             this.playlistEntryDatabaseService.addUpVote(playlistEntryId);
+            this.userDatabaseService.updateUserScore(entry.addedBy, -1);
             this.emitPlaylistUpdate(partyId);
-        } else if (userVotes.length === 1) {
-            // user has made a vote, if its a downvote, get rid of it
-            if (userVotes[0].type === PlaylistVoteEnum.UPVOTE) {
-                this.playlistEntryDatabaseService.removeUpVote(playlistEntryId);
-                this.playlistEntryDatabaseService.addDownVote(playlistEntryId);
-                this.playlistVoteDatabaseService.updateVote(playlistEntryId, userId, PlaylistVoteEnum.DOWNVOTE);
-                this.emitPlaylistUpdate(partyId);
-            }
+            this.emitUsersUpdate(partyId);
+        } else if (userVotes.length === 1 && userVotes[0].type === PlaylistVoteEnum.UPVOTE) {
+            this.playlistEntryDatabaseService.removeUpVote(playlistEntryId);
+            this.playlistEntryDatabaseService.addDownVote(playlistEntryId);
+            this.playlistVoteDatabaseService.deleteVote(playlistEntryId, userId);
+            this.playlistVoteDatabaseService.insertVote({
+                type: PlaylistVoteEnum.DOWNVOTE,
+                playlistEntryId,
+                userId
+            });
+            this.userDatabaseService.updateUserScore(entry.addedBy, -1);
+            this.emitPlaylistUpdate(partyId);
+            this.emitUsersUpdate(partyId);
         }
     }
 
-    public async addSongToPlaylist(partyId: string, userId: string, songId: string) {
+    public async addSongToPlaylist(partyId: string, userId: string, songId: string): Promise<boolean> {
         if (!this.playlistEntryDatabaseService.doesEntryExist(partyId, songId)) {
             const party = this.partyDatabaseService.getPartyById(partyId);
             const song = await this.spotifyService.getSpotifyAPI().tracks.getTrack(party.spotifyToken, songId);
@@ -190,9 +213,13 @@ export class PartyService {
                 playlistEntryId: entryId,
                 userId
             });
+            this.userDatabaseService.updateUserScore(userId, 1);
             this.emitPlaylistUpdate(partyId);
+            this.emitUsersUpdate(partyId);
+            return true;
         } else {
             //its already in the playlist
+            return false;
         }
     }
 
@@ -201,14 +228,26 @@ export class PartyService {
             partyId,
             PartyEvent.PLAYLIST_UPDATE,
             {
-                party: this.partyDatabaseService.getPartyById(partyId),
                 playlist: this.getPlaylistForPartyId(partyId)
             }
         );
     }
 
+    private emitUsersUpdate(partyId) {
+        this.eventEmitterService.emitEventAt(
+            partyId,
+            PartyEvent.USERS_UPDATE,
+            {
+                users: this.userDatabaseService.getUsersAtParty(partyId).sort(userSort)
+            }
+        );
+    }
+
     public removeSongFromPlaylist(partyId: string, userId: string, songId: string) {
-        // this.upNextService.emitEventToProcess(partyId, ProcessorEvents.PLAYLIST_REMOVE_SONG, {userId, songId});
+        if (this.playlistEntryDatabaseService.doesEntryExist(partyId, songId)) {
+            this.playlistEntryDatabaseService.removePlaylistEntry(partyId, userId, songId);
+            this.emitPlaylistUpdate(partyId);
+        }
     }
 
     public async search(partyId: string, searchTerm: string): Promise<any> {
@@ -252,5 +291,33 @@ export class PartyService {
             this.emitPlaylistUpdate(user.partyId);
             this.userDatabaseService.removeUserByTrackingId(trackingId);
         }
+    }
+
+    public async togglePlayback(partyId: string) {
+        const party = this.partyDatabaseService.getPartyById(partyId);
+        const partyState = this.upNextService.getPartyDataForPartyId(partyId);
+        if (partyState.isPlaying) {
+            await this.spotifyService.getSpotifyAPI().player.pause(party.spotifyToken);
+        } else {
+            await this.spotifyService.getSpotifyAPI().player.play(party.spotifyToken);
+        }
+    }
+
+    public async nextSong(partyId: string) {
+        const party = this.partyDatabaseService.getPartyById(partyId);
+        const playlist = this.playlistEntryDatabaseService.getAllPlaylistEntriesForParty(partyId);
+        if (playlist.length > 0) {
+            await this.upNextService.queueNextSong(playlist, party);
+            await this.spotifyService.getSpotifyAPI().player.nextSong(party.spotifyToken);
+        } else {
+            await this.spotifyService.getSpotifyAPI().player.nextSong(party.spotifyToken);
+        }
+    }
+
+    public async fixChromecastError(partyId: string) {
+        const party = this.partyDatabaseService.getPartyById(partyId);
+        const partyState = this.upNextService.getPartyDataForPartyId(partyId);
+        await this.spotifyService.getSpotifyAPI().player.addSongToEndOfQueue(party.spotifyToken, partyState.trackId);
+        await this.spotifyService.getSpotifyAPI().player.nextSong(party.spotifyToken);
     }
 }
